+ EntityDao implementation, dbunit dataset

+++ src/main/java/pl/chemik77/spring/addressbook/model/Address.java
@@ -17,7 +17,7 @@ public class Address implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	@Column(name = "street")
 	private String street;
@@ -34,11 +34,11 @@ public class Address implements Serializable {
 	public Address() {
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
